feat: generate typing for valid icon names and colors, update neutral color values
