+++ src/core/FeaturedLink.tsx
@@ -1,12 +1,13 @@
 import React, { CSSProperties, ReactNode } from "react";
 
 import Icon from "./Icon";
+import { ColorClass } from "./styles/colors/types";
 
 type FeaturedLinkProps = {
   url: string;
   children: ReactNode;
   textSize?: string;
-  iconColor?: string;
+  iconColor?: ColorClass;
   flush?: boolean;
   reverse?: boolean;
   additionalCSS?: string;
