+++ src/core/styles/properties.css
@@ -2,11 +2,11 @@
   :root {
     /* Neutral colors */
     --color-neutral-000: #ffffff;
-    --color-neutral-100: #f8fafc;
-    --color-neutral-200: #f4f8fb;
-    --color-neutral-300: #edf1f7;
+    --color-neutral-100: #f6f8fa;
+    --color-neutral-200: #eef1f6;
+    --color-neutral-300: #e6eaf0;
     --color-neutral-400: #e2e7ef;
-    --color-neutral-500: #c6ced9;
+    --color-neutral-500: #ced5de;
     --color-neutral-600: #adb6c2;
     --color-neutral-700: #89929f;
     --color-neutral-800: #667085;
